break out cloudscan components into different files
